pyproject.toml
[tool.poetry]
name = "midterm-project"
version = "0.1.0"
description = ""
authors = ["mexriddin maxkamtaev <maxkamtayev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^7.4.4"
pytest-check = "^2.2.3"
selenium = "^4.16.0"
appium-python-client = "^3.1.1"
faker = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = "tests"